+++ magda-web-client/src/helpers/popupUtils.ts
@@ -13,29 +13,8 @@ export const createPopupModeQueryString = () => {
     if (!params.size || !params.has("popup")) return "";
     const popupParams = new URLSearchParams();
     popupParams.append("popup", "true");
-    const datasetSubmitCallback = params.get("datasetSubmitCallback");
-    if (datasetSubmitCallback) {
-        popupParams.append("datasetSubmitCallback", datasetSubmitCallback);
-    }
     return popupParams.toString();
 };
-
-export function executePopupModeCallback(datasetId: string) {
-    const url = new URL(window.location.href);
-    const params = url.searchParams;
-    const datasetSubmitCallback = params.get("datasetSubmitCallback");
-    if (datasetSubmitCallback) {
-        const callback = window?.opener?.[datasetSubmitCallback];
-        if (callback && typeof callback === "function") {
-            // @ts-ignore
-            callback(datasetId);
-        } else {
-            throw new Error(
-                `Cannot locate dataset submission callback from opener window. Failed to notify the dataset submission event to the launcher system.`
-            );
-        }
-    }
-}
 
 export function useInPopUp() {
     const location = useLocation();